website/requirements.txt
django
djangorestframework
pillow==10.3.0